feat: moved h1cve to update security release
